Altered autonomous design.
