+++ Stronghold2016/src/com/nicoletfear/Stronghold2016/commands/AutonomousBackwards.java
@@ -1,46 +1,33 @@
-
 package com.nicoletfear.Stronghold2016.commands;
 
-import edu.wpi.first.wpilibj.Timer;
-import edu.wpi.first.wpilibj.command.Command;
+import edu.wpi.first.wpilibj.command.CommandGroup;
 
-import com.nicoletfear.Stronghold2016.Robot;
-import com.nicoletfear.Stronghold2016.subsystems.DriveTrain;
+public class AutonomousBackwards extends CommandGroup {
+	
+    public  AutonomousBackwards() {
+    	//addSequential(new Forward(5.5));
+    	addSequential(new Backwards());
+        
+        //backward = new Backward(2);
+        //addSequential(new Forward(2));
+    	//if (backward != null) backward.start();
+    	
+    	
+        // Add Commands here:
+        // e.g. addSequential(new Command1());
+        //      addSequential(new Command2());
+        // these will run in order.
 
-/**
- *
- */
-public class AutonomousBackwards extends Command {
+        // To run multiple commands at the same time,
+        // use addParallel()
+        // e.g. addParallel(new Command1());
+        //      addSequential(new Command2());
+        // Command1 and Command2 will run in parallel.
 
-    public AutonomousBackwards() {
-        // Use requires() here to declare subsystem dependencies
-        requires(Robot.driveTrain);
-        setTimeout(5);
-        //will have bot drive forwards for 5 seconds
-    }
-
-    // Called just before this Command runs the first time
-    protected void initialize() {
-    	Robot.driveTrain.drive(1 , 1);
-    }
-
-    // Called repeatedly when this Command is scheduled to run
-    protected void execute() {
-    }
-
-    // Make this return true when this Command no longer needs to run execute()
-    protected boolean isFinished() {
-        return isTimedOut();
-    }
-
-    // Called once after isFinished returns true
-    protected void end() {
-    	Robot.driveTrain.brake();
-    }
-
-    // Called when another command which requires one or more of the same
-    // subsystems is scheduled to run
-    protected void interrupted() {
-    	end();
+        // A command group will require all of the subsystems that each member
+        // would require.
+        // e.g. if Command1 requires chassis, and Command2 requires arm,
+        // a CommandGroup containing them would require both the chassis and the
+        // arm.
     }
 }
