+++ Stronghold2016/src/com/nicoletfear/Stronghold2016/commands/AutonomousForwards.java
@@ -1,47 +1,33 @@
-
 package com.nicoletfear.Stronghold2016.commands;
 
-import edu.wpi.first.wpilibj.Timer;
-import edu.wpi.first.wpilibj.command.Command;
+import edu.wpi.first.wpilibj.command.CommandGroup;
 
-import com.nicoletfear.Stronghold2016.Robot;
-import com.nicoletfear.Stronghold2016.subsystems.DriveTrain;
+public class AutonomousForwards extends CommandGroup {
+	
+    public  AutonomousForwards() {
+    	//addSequential(new Forward(5.5));
+    	addSequential(new Forwards());
+        
+        //backward = new Backward(2);
+        //addSequential(new Forward(2));
+    	//if (backward != null) backward.start();
+    	
+    	
+        // Add Commands here:
+        // e.g. addSequential(new Command1());
+        //      addSequential(new Command2());
+        // these will run in order.
 
-/**
- *
- */
-public class AutonomousForwards extends Command {
+        // To run multiple commands at the same time,
+        // use addParallel()
+        // e.g. addParallel(new Command1());
+        //      addSequential(new Command2());
+        // Command1 and Command2 will run in parallel.
 
-    public AutonomousForwards() {
-        // Use requires() here to declare subsystem dependencies
-        requires(Robot.driveTrain);
-        //makes bot drive forwards for five seconds
-        setTimeout(5);
-    }
-
-    // Called just before this Command runs the first time
-    protected void initialize() {
-    	Robot.driveTrain.drive(-1 , -1);
-    	//values inverted because it was driving backwards for some reason
-    }
-
-    // Called repeatedly when this Command is scheduled to run
-    protected void execute() {
-    }
-
-    // Make this return true when this Command no longer needs to run execute()
-    protected boolean isFinished() {
-        return isTimedOut();
-    }
-
-    // Called once after isFinished returns true
-    protected void end() {
-    	Robot.driveTrain.brake();
-    }
-
-    // Called when another command which requires one or more of the same
-    // subsystems is scheduled to run
-    protected void interrupted() {
-    	end();
+        // A command group will require all of the subsystems that each member
+        // would require.
+        // e.g. if Command1 requires chassis, and Command2 requires arm,
+        // a CommandGroup containing them would require both the chassis and the
+        // arm.
     }
 }
